Addition of setting 92 (watt %) including stable default value
Added setting 92 to the driver and config of the Smart Dimmer 6 to
prevent overkill of power reports; factory default setting is 5%
resulting in multiple reports per second depending on the device
connected to it.

For other devices, like Fibaro, Neo, Greenwave, this threshold has
already been raised to 80%. If needed, the user is able to lower the
value (consciously). App default value for the ZW099 changed to 80%.

In addition, I recommend to change the app default setting of parameter
91 back to 25 W (factory default).

I was not able to test this setting myself due to missing dimmer-6;
pretty straightforward though (size-1 parameter, range 0 - 100)